Set the width of the wrapper to 800px and set the left and right margins to auto.
This is a very common page layout strategy for print designers who come to the web. They fix the width of the content and ignore the width of the browser window.

Note that the line length, or measure, is too long for comfortable reading in this example (as is the leading).